requirements.txt
asyncio==4.0.0
fastapi==0.121.3
fastapi-mail==1.5.8
uvicorn==0.38.0
pydantic==2.12.4
sqlmodel==0.0.27
sqlalchemy==2.0.44
asyncpg==0.30.0
python-dotenv==1.2.1